+++ FamilyTreeLib/src/Repositories/IPersonRepository.cs
@@ -5,6 +5,6 @@ namespace FamilyTreeLib.Repositories;
 public interface IPersonRepository {
   Task InsertAsync(Person person);
   Task UpdateAsync(Person person);
-  Task<Person?> GetByIdAsync(long id);
+  Task<Person?> GetByIdAsync(int id);
 
 }
